+++ kitty_tests/screen.py
@@ -280,6 +280,35 @@ class TestScreen(BaseTest):
         self.ae(s.cursor.x, 1)
 
     def test_resize(self):
+        from kitty.window import as_text
+        def at():
+            return as_text(s, add_history=True)
+        # test that a wrapped line split by the history buffer is re-stitched
+        s = self.create_screen(cols=4, lines=4, scrollback=4)
+        text = ''
+        for i in range(s.lines + 1):
+            text += str(i + 1) * s.columns
+        s.draw(text)
+        self.assertTrue(s.historybuf.endswith_wrap())
+        s.resize(s.lines, s.columns + 2)
+        self.assertTrue(s.historybuf.endswith_wrap())
+        self.ae(str(s.historybuf), '111122')
+        self.ae(at(), text + '\n')
+        s = self.create_screen(cols=4, lines=4, scrollback=4)
+        s.draw('1111222'), s.linefeed(), s.carriage_return()
+        s.draw('333344445555')
+        s.resize(s.lines, s.columns + 2)
+        self.ae(str(s.historybuf), '111122')
+        self.ae(str(s.line(0)), '2')
+        self.ae(at(), '1111222\n333344445555\n')
+        s = self.create_screen(cols=4, lines=4, scrollback=4)
+        s.draw('1111😸2'), s.linefeed(), s.carriage_return()
+        s.index(), s.index()
+        s.resize(s.lines, s.columns + 1)
+        self.ae(str(s.historybuf), '1111')
+        self.assertTrue(s.historybuf.endswith_wrap())
+        self.ae(at(), '1111😸2\n\n\n')
+
         # test that trailing blank line is preserved on resize
         s = self.create_screen(cols=5, lines=5, scrollback=15)
         for i in range(3):